+++ base_accounting_kit/doc/changelog.md
@@ -64,3 +64,32 @@
 #### Version 13.0.4.5.6
 #### UPDT
 - Code Optimized-dashboard
+
+#### 24.04.2020
+#### Version 13.0.4.6.7
+#### IMP
+- Multi Company issue updated.
+
+#### 29.04.2020
+#### Version 13.0.4.7.8
+#### UPDT
+- Report updated.
+
+#### 10.06.2020
+#### Version 13.0.4.8.9
+#### UPDT
+- Comma separator in Dashboard
+- Currency symbol position in Dashboard
+- Monthly invoice filter in Dashboard
+- Check payment issue, in case of multi payment
+
+
+#### 16.07.2020
+#### Version 13.0.4.8.10
+#### UPDT
+- Multi Company in Dashboard
+
+#### 10.09.2020
+#### Version 13.0.4.8.11
+#### UPDT
+- Credit note, refund removed from dashboard